+++ tproxy_other.go
@@ -8,7 +8,7 @@ import (
 	"github.com/e1732a364fed/v2ray_simple/utils"
 )
 
-func ListenTproxy(addr string) (tm *tproxy.Machine) {
+func ListenTproxy(addr string) (_ *tproxy.Machine) {
 	utils.Warn("Tproxy not possible on non-linux device")
 	return
 }
