+++ utils/utils.go
@@ -3,18 +3,10 @@ package utils
 
 import (
 	"flag"
-	"math/rand"
 	"strings"
-	"time"
 
 	"github.com/BurntSushi/toml"
 )
-
-func init() {
-	//保证我们随机种子每次都不一样, 这样可以保证go test中使用随机端口时, 不同的test会使用不同的端口, 防止端口冲突
-	// 因为我们所有的包应该都引用了 utils包, 所以可以保证这一点.
-	rand.Seed(time.Now().Unix())
-}
 
 // bufio.Reader 和 bytes.Buffer 都实现了 ByteReader
 type ByteReader interface {
